Fix JS locale data for release of Money 6.19
The money gem in its latest iteration added a space between value and
unit of the Swiss Franc. Our heuristic in `_js_locale_data.html.erb`
can't deal with that, so we add a third branch to it: If the money gem
has a format wish, we'll respect it, but replace its abbreviations with
the one the `accounting.js` library prefers.

(cherry picked from commit f19e6fcab59db4887ba46c36ee6e2616c06d3a46)