Products that are not available shouldn't show Available status
Currently if a product has the `available_on` date in the future, it
will still show a status of Available, even though it is not available
for purchase on the store. This change proposes a new status of
`Unavailable` for scheduled products.

Co-authored-by: Harmony Bouvier <harmony@super.gd>
(cherry picked from commit e404e448705af3e166a5197ff1c4f0fbf0d984af)