Allow turning solidus_admin on and off for certain specs
We need a way of a spec telling the app to use the new admin for tests
that require it, and using the old admin to not use it. After
experimenting with using a cookie, which led to a number of undesirable
issues (needing a new gem to manage cookies across Capybara drivers,
having to visit some URL before being able to even set the cookie etc),
what we do is we just define a global module attribute accessor on the
dummy app that enables or disables the routes for the `solidus_admin`
gem.