Properly delegate Spree::ApiConfiguration#promotion_attributes
Initializing the app with the code prior to this commit would
instantiate the promotions configuration, which is not good because
instantiating it via `Spree::Config.promotions` leads to the object
being instacached and very, very hard to change.

Also, we currently have to spots in the app where we can configure the
promotion api attributes, and this reduces it to just one by deprecating
the preference :promotion_attributes on `Spree::ApiConfiguration`.