Provide formatted date values to field component
This change fixes an issue where certain browsers (namely Firefox) will
not render the date field correctly when it contains the full date and
time format as the value. This change just truncates the available_on
and discontinue_on values to match the precision expected, rather switch
to the datetime-local input which is not fully supported in all
browsers, but also maintains the old admin behaviour.

Co-authored-by: Andrew Stewart <andrew@super.gd>