Configure promotions via a configuration instance
Prior to this, we would only allow changing the configuration of the
promotion system by switching the `promotions_configuration_class`. The
`AppConfiguration` would then instantiate that class, and one would be
able to do all the necessary customizations via the instance.

This allows customizing a promotion configuration and then setting it as
the promotion configuration class. It's a lot easier to understand.