Add a deprecated configurable class
This class will act as a stand-in for those classes that are currently
configurable via Solidus' `AppConfiguration` class, but are only used
from *within* the legacy promotion system. We have delegated these
extension points to `Spree::Config.promotions`, so the
`Spree::Core::NullPromotionConfiguration` needs to implement methods for
them, but I don't see a lot of sense in providing Null implementations
for things that aren't even needed.