Rename PromotionConfiguration to LegacyPromotionConfiguration
Earlier in the extraction process, I've defined a configuration object
for the current promotion system and called it
`Spree::Core::PromotionConfiguration`. However, it is much more aptly
named `SolidusLegacyPromotions::Configuration`, as it will be moved into
that gem shortly. This way it's also clear when we move to the
`Spree::Core::NullPromotionConfiguration` what we are changing.