Backend: Make order search fields configurable
The order search depends on legacy promotion system associations to be
present, but we don't want to introduce Deface as a core dependency.
Therefore, we must make the search fields for orders configurable.

This introduces a new configuration option on the Backend configuration:
`Spree::Backend::Config.search_fields`. This Hash has controller paths
as keys, and arrays that holds hashes with a `:partial` key that indicates
which partial should render this search field.

Optionally, one can specify an `:if` key with a lambda as the value that
gets executed at runtime, so that unnecessary fields ("Store" for
installations with only one store instance) are not displayed.

Complex fields can be done by specifying a custom partial, such as the
`:date_range_picker` partial in this PR.

The third key, `:locals` contains a has of the local variables inside
the partial. Anything that needs to be computed dynamically should be
done with a lambda here.