Create custom orders index component for solidus_legacy_promotions
The new admin allows orders to be filtered by promotion name. However,
with the legacy promotion system extracted into an extension, we can't
rely on `Spree::Promotion` to exist in `solidus_admin`. In order to get
around this issue, we create a new index component for the admin orders
page that includes the filter, and register it in an initializer.

This also adds a spec for filtering by promotion name.