Allow safe reloads of the component registry
This adds a new `ComponentRegistry` class that allows:
- setting component classes as strings (to allow for easy reloading)
- getting nice errors when requesting components that don't exist
- getting nice errors when setting a component to a non-existing class

It also moves some of the logic from `SolidusAdmin::Configuration` into
a separate class.

I've decided not to add memoization here as I'm not sure this would make
a giant difference performance-wise, and also the Rails Guide on
Reloading constants very clearly says:

```
Bottom line: do not cache reloadable classes or modules.
```