Use configurable adjustment promotion source types in Thumbnail.for
There's a View Component for rendering icons, and it's got a convenience
method for rendering adjustables to adjustments.

This uses the `Spree::Config.adjustment_promotion_source_types`
configuration option to figure out whether an adjustable needs a
megaphone as an icon.

With the legacy promotion system extracted, `Spree::PromotionAction` will not
necessarily be available.