Patch Order State Machine rather than order
Some extensions rely on the default state machine being fully
reloadable, and as long as we use solidus_legacy_promotions, the
before_transition method `:apply_shipping_promotions` must survive a
reload of the state machine.

Rather than patching Spree::Order and adding the transition here, we add
it where all the other default transitions are added: In the class
methods of the order state machine.