Add before action to handle option type params
As reported in #5751, the new admin UI has a bug with the option type
selectors: only one option type can be added.

We narrowed this down to an issue handling option type params in the new
admin products controller. We noticed legacy controller contains a
before action to handle the incoming option type params, but the new one
did not. Re-adding the before action solves the bug. This is required
because the form input for option types contains comma separated values
within a string:

`<input value="1,2" type="hidden">`

Co-authored-by: An Stewart <andrew@super.gd>
Co-authored-by: Kendra Chateau <kendra@super.gd>