Add solidus_user_roles migrations to core
These migrations originally come from `solidus_user_roles`. We intend to
migrate some of the functionality from that extension into `core`, and
this is the first step. With the table check, these migrations should be
safe to (re)run for any store which might already have
`solidus_user_roles` installed.