Remove inaccessible details/summary element
The variant search on the order#show page utilized a `details`/`summary`
HTML element in order to get open/close functionality. However, Axe
flags this as inaccessible, because the `summary` element does not have a
label. This removes the `summary` element in favor of a simple div, and
changes the controller to rely on a `hidden` class instead.