Add new CheckboxRow admin component
While I am not sure how frequently this will be re-used, the fact that
it comes up 14 times across the roles/new and roles/edit pages makes
this extraction worthwhile. I also tried out a partial and a helper
method previously but both approaches were messy and poorly architected.
This should serve us much better in the long run and be easier to
maintain.