Appease Codecov with new model specs
I would not normally write specs to ensure associations, as that seems
like it's not my responsibility to test, but Codecov seems to think
otherwise and would like a spec for `Spree::RolePermission`, a model
with nothing on it except some `belongs_to`s. I also added specs for the
other models that were flagged.

The association specs are a bit uglier than they could be if we were
using the `shoulda_matchers` and their `is_expected.to belong_to` but it
seems like it's not worth it to install it just to beautify 3 specs!