Add privilege & category to PermissionSets
Later commits proved the need for an easier method of identifying and
categorizing permission sets more cleanly on the front-end, which is why
we added the category & privilege columns. Each subclass (eg.
`Spree::PermissionSets::ConfigurationDisplay`) has its own privilege and
category, which is then used to build the instance
(Spree::PermissionSet) with the correct identifying features.