Change default scope in Users admin
I feel pretty strongly that the default scope should be "all" instead of
"customers". That way, when presented with a long list the first thought
is to filter it, instead of the first thought being "why is this user
missing" and then realizing that it's because it's filtering by users
with the "customer" role only.