Display last_sign_in_at instead of created_at
The new users admin interface now shows the last sign in date instead of
the date the user was created. It will range between "Never", "Less than
a minute ago", "3 days ago", "2 months ago" etc. It uses the built-in
logic provided by `ActionView::Helpers::DateHelper#time_ago_in_words`.

The specs are a bit tricky and stubby due to the need to keep them
general and agnostic of whatever is being used for the
`Spree.user_class`.