Variant Autocomplete: Allow passing select2 options
This change allows users of the variant autocomplete to customize the
behavior of select2, rather that just discarding all the keys that are
not `searchParameters`.

We're using this for a custom controller in which we want to be able to
filter by variant, and be able to clear the selection.

So in that controller, we can do:

```js
const inputElement = document.getElementById('variantInput')
inputElement.variantAutocomplete({ allowClear: true })
```