Add new users admin addresses page
This migrates the `users/:id/addresses` page to the new admin. Instead
of squashing everything into one controller action with an `if
request.put?` wrapper (like the legacy backend controller did), I chose
to break this new admin into two actions separated by their methods and
responsibilities.

The old admin page did not have anything in the way of validations or
error messages, it would simply fail to update the address and leave it
as is if the user provided invalid values. (While also showing a message
saying that the update was successful?)

This page improves upon that flow, adding validations to the address
forms when their values are invalid and preventing the update, but
since we are submitting an update against the user and not the address
directly, we still need to do some extra work to run the validations and
generate errors for the form (rather than silently failing to update
like the legacy back-end controller used to do).