Introduce concept of template and non_template variants
All products have variants, but not all variants can actually be sold.
When a product has option types, the sellable variants are those that
are assigned to option values, whereas the "master" variant is only
there as a template for other variants to be created.

I'm using the language "template variant" because "master variant"
because it more closely matches what the `is_master` flag on the
Spree::Variant actually does.