Handle states_required? in admin address component
Previously, this component would just disable the state_id field when
the selected country did not require states, but this meant that the
user would never be able to self-fill the state_name field, which is
something that some stores can require with the
`Spree::Config[:address_requires_state]` preference.

The component now takes this into account and will swap between a
state_id dropdown select field, or a blank state_name text input
depending on the country the user has selected.

This brings this new admin component up to par with the previous backend
address forms.