Extract user stats into shared component
This DRYs things up a bit so that all pages of the users admin can reuse
the same helper, translations, and component. Since the order index does
not use this stats component, but does make use of the last_login logic,
that was broken out into a separate helper for ease of use.