Render variant shipping category
If a variant does not have a shipping category defined, the form will
render with the shipping category from the associated product. If you
submit the form, the variant will now have the same shipping category as
associated the product.

We fix this by always rendering the form with the shipping category
associated with the variant.

Co-authored-by: An Stewart <andrew@super.gd>