Do not use solidus_frontend to test extension test app generator
solidus_frontend is [EOL since 2024-10-21](https://github.com/solidusio/solidus_frontend?tab=readme-ov-file#-warning)
using it to test the extension:test_app rake task is not working since
it is not compatible with latest Rails, Ruby and Solidus.

Since we already have a test for the extension test app generator we
can check for it's dummy app existence instead.