Backend Menu: Fix "Promotions" items
Prior to this commit, legacy promotions would be called "Promotions",
but if `solidus_promotions` is installed, they would be called "Legacy
Promotions" in the menu. It's much easier if each of the two promotion
engines just inserts its own menu item into the menu, and the legacy
promotion system is called "Legacy Promotions" from the get-go.

Co-Authored-By: thomas@vondeyen.com