Fix component translation scopes
In the `view_component` gem, the `i18n_scope` for a component depends on
the file name of the component, and removes `app_components` and
everything before that in order to construct a scope.

This fails to work with components that live in `lib/components/admin/`,
as with our extensions, as there is no `app/components` in the path to
these components.

This change uses the class name instead, which should be easier on
memory and correctly works with components from solidus extensions.

This also fixes translations in inherited components, allowing to delete
some redundant translations in the promotion gem's `order` component.