Fix(Admin menus): Stable promotion menu items
This makes sure that the promotion menus are always below the product
menu item, in both new admin and the backend.

It also changes the nomenclature to be "Promotions" for the legacy
promotion system and "Promotions (new)" for the new promotion system.

(cherry picked from commit 6d755154c06968f1b49d63979a35fe22218a0b19)