Fix admin promotions controller
This controller still did not know how what to authorize against, and
wanted to visit a URL that doesn't exist when clicking on a promotion.

This also changes at least the name of each promotion to be a link
element that can easily be targeted with Capybara and works with all
major browsers.

(cherry picked from commit 2eccd6d6d5b4eaeaad2c3ace32409ea87d821443)