Use semantic UI elements in admin table rows
All over the new admin we use divs or simply text to display the
resource and rely on a complex Stimulus click event handler to open the
edit route. This is not semantic at all and makes it necessary to use a
fully JS enabled browser under all circumstances to test the new admin
interface which leads to flaky and slow tests.

We should use links instead and probably remove the stimulus click event
handler. Also, links are not visually highlighted, which leads to a11y
issues. So, this commit adds underlines to links in table rows.

https://github.com/solidusio/solidus/issues/5944