Convert symbol columns to links in admin tables
Previously, when provided a symbol, the admin table component would
simply call `data.public_send(cell)` (with cell being the symbol). This
would result in a basic string as the content of the table row.

We want to stop relying on the additional Stimulus click handlers
that captured those plain text clicks and redirected the users to
wherever the `row_url` location was.

Now, the `render_data_cell` method handles wrapping the plain text in a
link, with the href being the `row_url` location. Symbols can still be
provided as columns and the component takes care of generating the proper
semantic UI elements for us.