Backend: Add missing error translation
We were missing a translation here, but never noticed, because we were
not testing for an actual human-understandable string.

This is broken in the v4.0, v4.1, v4.2 and v4.3 branches with a noisier
error message, because in those branches, `I18n.t("unavailable")` is
different from `t("unavailable")`.

(cherry picked from commit 506ddfcbd34703c00dbb75b5197646f70099f726)