Fix(promotions): Add model_class method to PromotionCodesController
Because the new promotion system lives outside the `spree` namespace, we
need to be diligent about adding `model_class` to all the controllers.

I have not added a spec, because this only fails if
solidus_legacy_promotions is not loaded, and that is not the case in our
test suite.