Fix(v4.1): Collapse admin nav on small windows
It is not clear to me why this only fails on some builds, but not
others. However, Capybara/Selenium give us a screen size of 800, and a
really large browser window of 1920. Let's check for the browser size
instead of the screen size, as I've not found a good way of fixing the
screen of a Selenium/Capybara session.