Utilize variant tax category when calculating taxes if possible
In the default tax_calculator, this change ensures that the variant’s
current tax_category is considered when calculating taxes.

In the case of the item_rates method, only line_item items will have
this method, so it is accessed with `try` first, with a fallback to
without `variant_`. `try` is necessary to use instead of `.&` or
`try!` because `Spree::Shipment`s and `Spree::ShippingRate`s do not
implement `#variant_tax_category_id`.

Co-authored-by: Adam Mueller <adam@super.gd>