Allow select component to accept non-object fields
Sometimes, as part of a form, you might want to include a field for
something which is not strictly part of the object you are updating.

In this instance it was a form for updating a StoreCredit, which needed
to include a select for a StoreCreditReason id. Since StoreCreditReason
is only linked to StoreCredit via a StoreCreditEvent, this select was
blowing up when we tried to use it (due to trying to call
store_credit.public_send(store_credit_reason_id) which is not defined).

This extra check allows this form field to be used more freely and
flexibly without immediately blowing up.