Replace unauthorized_redirect lambda with configurable class
This lambda has to be set on the instantiated class, requiring the
controller it's set on to be autoloaded on boot. Loading a controller
also loads all helpers, and that takes quite a bit of time.

So this comes with a performance benefit, because neither the controller
nor the handler need to be loaded at configuration time.