[SolidusAdmin] Fix mock_component helper
The mock component in the ComponentHelpers module was not a real
constant, which messes with ViewComponent's expectations about what
render_inline is given. Using `stub_const` in the helper allows us to
give it an actual name, and view_component > 3.21.0 will work for us.

The helper is only used in the base component spec. I could have spent
more time giving it an optional block, but this solution is the most
straightforward.

(cherry picked from commit 228ca124326073fb88bf789416a65a13e664cb9b)

# Conflicts:
#	admin/lib/solidus_admin/testing_support/component_helpers.rb
#	admin/spec/components/solidus_admin/base_component_spec.rb