Set tax adjustment amount without persistance
To support the in-memory order updater, we want adjustments to be
modified without writing to the database immediately.

We added `autosave: true` in the association on line_items ensures that
the adjustment amount will be saved with the line_item, or by extension
the order when it is eventually persisted by the order updater.

The tests for the OrderTaxation have been updated to use `size` instead
of `count` to ensure an in-memory reported count/size.

Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Andrew Stewart <andrew@super.gd>
Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>
Co-authored-by: benjamin wil <benjamin@super.gd>
Co-authored-by: Alistair Norman <alistair@super.gd>