Mark adjustments for removal in `OrderTaxation`
Instead of actually destroying adjustments in the OrderTaxation class,
we can just mark them for removal. This will allow them to be removed
when the order is persisted, instead of being removed during the
recalculate process. This is an improvement in service of recalculating
order changes without writing to the database.

We also need to ensure `Order#adjustments` association has autosave
enabled so that the marked for destruction adjustments are removed when
we persist the order.

We're also improving our OrderUpdater tests to better verify our changes
maintain the existing behaviour.

Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Alistair Norman <alistair@super.gd>
Co-authored-by: Chris Todorov <chris@super.gd>
Co-authored-by: Harmony Bouvier <harmony@super.gd>
Co-authored-by: Kendra Riga <kendra@super.gd>
Co-authored-by: Senem Soy <senem@super.gd>
Co-authored-by: Sofia Besenski <sofia@super.gd>
Co-authored-by: benjamin wil <benjamin@super.gd>