Refactor `ui/forms/address` component
Changes description:
 - introduced `@include_name_field` flag that controls whether to show
 "Name" field on the form;
 - renamed keyword `address` to `addressable` to better reflect that the
 nature of the passed object;
 - renamed keyword `name` to `fieldset_name` to avoid confusion with
 "Name" field, whereas `fieldset_name` is a name to be used for the form (e.g. "users[billing_address_attributes]");

Motivation:
This change allows to reuse this component in location stock form
component, where the only difference between forms is that "Name" field
is not there for location stock address.

Also, maybe it'll be sensible in the future to migrate stock location to
use a `stock_location_address` association and remove all the address
related fields from the `stock_location` table, but for now it's quite a
big task to be contained in current scope. So for now, just treat
`stock_location` as an `addressable` which can be used to render address
form component.