Refactor ui/forms/checkbox component to include common form elements
Motivation:
We have a couple of places throughout the admin where checkboxes
are used as part of a form, with all the same elements being reused
 to construct a checkbox field:
 - label with some caption;
 - hidden input field to support sending a "0" value in request;
 - optional toggletip hint;
Few places however lacked the hidden input, therefore the form was
not submitting the "0" value correctly.

This updated component builds a reusable form checkbox, that
encapsulates all the needed layout and logic for a form checkbox to
be displayed and work properly, with an interface to customise it
to each form needs.