Use `human_attribute_name` for promo calculator labels
Promotion calculators have preferences, and these need to be translated
with the correct scope.

When you specify a preference for a calculator, a virtual attribute is
generated. For example:

```
class MyCalculator < Spree::Calculator
  preference :percent, :integer, default: 10
end
```

In this scenario, we get a virtual attribute `preferred_percent` on the
`MyCalculator` class. The expectation is that then one can translated
that attribute by calling:
`MyCalculator.human_attribute_name(:preferred_percent)`, and - crucially
- that this attribute is used in form labels.

While all other calculator forms use this (through `form.label
:preferred_percent`) or similar, the default form did not, and instead
used `I18n.t(:percent, scope: :spree)`, which is not specific to the
calculator, and might lead to mistranslations.