Use Firefox for system specs
Chrome seems to consistently get updated with new interesting options,
and our specs have been sometimes more, and sometimes less flaky with
Chrome as the browser that drives our system specs.

This change uses the preconfigured `selenium_headless` driver from the
Selenium gem, which uses Firefox under the hood.

There are a few changes to helpers in the new admin. All of them have to
with clicking table cells rather than the underlying table row. This is
how real users also interact with the respective page, so I think it's a
good change.