Allow option types sorting
Sortable URL parameter, which is required for sorting to
work properly, was not present for option types table after
changing the index table to use explicit links rather than
`row_url`s, due to condition in the table component
template. Here we move `data-sortable-url` attribute out of
the condition, so it's applicable to any index page that
defines `sortable_options`.