Add in-memory updater patch for legacy promotion system
This change introduces some high-level integration tests and a patch
similar to the one for the regular order updater, which provides a
compatibility layer for the legacy promotion system, specifically around
the eligible flag on adjustments.

We want to protect against manipulative database queries in the legacy
promotion system. We also want to ensure that objects in memory have
their attributes changed correctly but not persisted.

Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: Kendra Riga <kendra@super.gd>
Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Alistair Norman <alistair@super.gd>
Co-authored-by: Senem Soy <senem@super.gd>
Co-authored-by: Sofia Besenski <sofia@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>