Add 500ms delay before AJAX in Select2
We have four select2 Ajax pickers in `solidus_backend`:
- Option Value Picker
- Product Picker
- Taxon Autocomplete
- Variant Autocomplete

The Variant autocomplete already has a 500ms delay between the user
finishing typing and the Ajax request firing. This is good, because it
stops requests happening at every keypress. This commit extends that
behavior to the option value, product and taxon pickers.