Create a manipulative query handler
This will be used by the in-memory order updater to log manipulative
queries when the persist flag is set to false. This will inform
application developers whether their application is complying with the
in-memory order updater's "no writes" philosophy.

Previously the `db-query-matchers` gem we use to check for manipulative
queries was only used in tests. However, we now use it in application
code so the dependency must be moved to the gemspec.

Co-authored-by: Alistair Norman <alistair@super.gd>
Co-authored-by: Sofia Besenski <sofia@super.gd>
Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Chris Todorov <chris@super.gd>
Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: benjamin wil <benjamin@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>