Improve transitioning from batch mode
Having "search" mode as default transition from "batch" mode was not
ideal because if a table is sortable and does not have a search field,
leaving "batch" mode would turn on "search" mode, mode which disables
sorting.

Instead, let's just remember which mode we were on before going into
"batch", and then restore it.