Country: Add dependent/inverse_of options
This also removes a bogus spec that tested admin behavior when dealing
with the situation where a country was removed that had been used for an
address. This commit prevents that entire situation.

The original issue that was solved with the PR that introduced the spec
was https://github.com/spree/spree/issues/3571 - allowing admins to
add/remove countries, but even here the original developers did not deal
with dependent records. For the use case described there - i would like
to change the countries my store services - I think using zones is a
better solution that deleting the country.